pvr: add an inline method OpenPVRDatabase() that tries to open the db and logs an error when it fails

+++ xbmc/pvr/dialogs/GUIDialogPVRChannelManager.cpp
@@ -453,10 +453,11 @@ bool CGUIDialogPVRChannelManager::OnMessage(CGUIMessage& message)
         {
           if (pItem->GetPropertyBOOL("Virtual"))
           {
-            CPVRDatabase *database = g_PVRManager.GetTVDatabase();
-            database->Open();
-            database->Delete(*pItem->GetPVRChannelInfoTag());
-            database->Close();
+            if (CPVRDatabase *database = OpenPVRDatabase())
+            {
+              database->Delete(*pItem->GetPVRChannelInfoTag());
+              database->Close();
+            }
 
             m_channelItems->Remove(m_iSelected);
             m_viewControl.SetItems(*m_channelItems);
@@ -509,12 +510,14 @@ bool CGUIDialogPVRChannelManager::OnMessage(CGUIMessage& message)
                 newchannel.SetStreamURL(strURL);
                 newchannel.SetClientID(XBMC_VIRTUAL_CLIENTID);
 
-                CPVRDatabase *database = g_PVRManager.GetTVDatabase();
-                database->Open();
+                CPVRDatabase *database = OpenPVRDatabase();
+                if (!database)
+                  return false;
+
                 database->Persist(newchannel);
                 database->Close();
-                CFileItemPtr channel(new CFileItem(newchannel));
 
+                CFileItemPtr channel(new CFileItem(newchannel));
                 if (channel)
                 {
                   channel->SetProperty("ActiveChannel", true);
@@ -716,8 +719,8 @@ void CGUIDialogPVRChannelManager::SaveList() // XXX investigate: renumbering doe
   if (!m_bContainsChanges)
    return;
 
-  CPVRDatabase *database = g_PVRManager.GetTVDatabase();
-  if (!database || !database->Open())
+  CPVRDatabase *database = OpenPVRDatabase();
+  if (!database)
     return;
 
   CGUIDialogProgress* pDlgProgress = (CGUIDialogProgress*)g_windowManager.GetWindow(WINDOW_DIALOG_PROGRESS);
